+++ app.rb
@@ -1,16 +1,19 @@
 require 'slim'
 require 'json'
-require_relative 'app/models/screaming_frog'
+require_relative 'app/models/url_stats'
 
 require 'pry'
 require 'pry-nav'
 
-get "/" do
-  slim :index
-end
 
 post '/work' do
   params = JSON.parse(request.env["rack.input"].read)
   content_type :json
-  ScreamingFrog.process(params['url'], '/')
+  payload = {
+    sensor_id: params['sensor_id'],
+    run_token: params['run_token'],
+    data: {}
+  }
+  params['urls'].each { |u| payload[:data][u] = UrlStats.process(u) }
+  payload.to_json
 end
